Added a new built-in variable '_MyName' which can be set to give a character a custom name. A lot of other changes were required along with it.
Speech log uses its own data structure now, which includes the custom name that can be used to overwrite the name displayed in the log. This required changing all of the usages of that vector.

Code for populating speech log was also modified

UI changes necessary to power the whole thing