A big rewrite of FaceWidget.
Why: I was chasing a bug, where the game would not switch back from custom image avatar to a built-in avatar if you changed player roles. Then when I fixed it something else broke. And then I fix what broke and yet another thing broke. Then I realized the way FaceWidget is written and used is flawed in two ways.

Flaw 1 - trying to fit as much state in as few variables. There was the whole thing with locking moods which was made to not have to have separate set of variables for speakers, which was the main source of issues.

Flaw 2 - a lot of logic that belonged in `FaceWidget` was spread around `GameScreen` , specifically figuring out how to manipulate the moods and the locks.

Analysis - when you think about it, there are essentially three differnt things displayed in the widget. Player's face, Speaker's face and Dying face. Technically Dying face is Player's face but because it behaves slightly differently (it has no special state beyond 'I am active') I think it makes sense to keep it separate to avoid issue.

Solution - there is a new struct that holds Face's state, and all the logic in `GameScreen` that previously had to do weird stuff to figure out how to manipulate the state to get the desired effect just directly tells the widget what to do and which face layer to modify. Speech on start enables Speech face, and disables it when speech ends. Ditto for dying. I think the solution is much more elegant, though there are likely ways to improve it still.